requirements.txt
black==24.10.0
build==1.2.2.post1
CacheControl==0.14.0
camelot-py==0.11.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
cleo==2.1.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
crashtest==0.4.1
cryptography==44.0.0
cycler==0.12.1
distlib==0.3.9
dulwich==0.21.7
et_xmlfile==2.0.0
fastjsonschema==2.20.0
filelock==3.16.1
fonttools==4.55.0
idna==3.10
installer==0.7.0
isort==5.13.2
jaraco.classes==3.4.0
keyring==24.3.1
kiwisolver==1.4.7
matplotlib==3.9.3
more-itertools==10.5.0
msgpack==1.1.0
mypy-extensions==1.0.0
numpy==2.1.3
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
pdfminer.six==20240706
pexpect==4.9.0
pillow==11.0.0
pkginfo==1.11.2
platformdirs==4.3.6
poetry==1.8.4
poetry-core==1.9.1
poetry-plugin-export==1.8.0
ptyprocess==0.7.0
pycparser==2.22
pyodbc==5.2.0
pyparsing==3.2.0
pypdf==5.1.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32-ctypes==0.2.3
RapidFuzz==3.10.1
requests==2.32.3
requests-toolbelt==1.0.0
shellingham==1.5.4
six==1.16.0
SQLAlchemy==2.0.36
tabulate==0.9.0
tomlkit==0.13.2
trove-classifiers==2024.10.21.16
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.27.0
